Lotto 클래스 내부 로또 숫자의 길이 유효성 검사 기능 구현
